OPEN - issue ROO-9: Identical field name with entity name causes scaffolded JSP to not render correctly http://jira.springframework.org/browse/ROO-9
This is due to Spring MVC taglibs auto generating 'id' attributes which are the same as the specified 'path' and 'modelAttribute'.

So

modelAttribute="hello" and path="hello" will both create id="hello" which leads to invalid XML / html which in turn causes the Browser to not render correctly.

This is fixed now by explicitly generating id="_hello"