ROO-210: Dateformat conversion for free... as yyyy-MM-dd
Added a new dateFormat parameter to the @WebScaffold annotation. This allows you to define the date format on a per-controller basis and is reflected in the initbinder method of the controller itself as well as its related view artifacts (JSPs and Dojo artifacts).

The automatic controller command reflects this with the new (optional) -dateFormat flag:

new controller automatic -name ~.web.OwnerController -formBackingObject ~.domain.Owner -dateFormat dd-MM-yyyy