ROO-3311: Provide option for Spring Roo managed XML configuration for Services, Controllers, GWT Locators, etc.
Added ability to instantiate services using XML configurations.  Added a propert to RooService (useXmlConfiguration) to control this behavior.   This is the default method when using GAE.  Changed ServiceLayerProvider to create and autowire a set method rather than a field to facilitate XML configurations.  Began modifying GwtServiceLocatorImpl to provide the capability to instantiate Locators using XML configuration.